chore: address batch review for issues #605-#609
