+++ app/services/refraction_static_artifacts/contract.py
@@ -5,35 +5,12 @@ from __future__ import annotations
 from dataclasses import dataclass
 from pathlib import Path
 
-from app.services.refraction_static_design_matrix import (
-    REFRACTION_DESIGN_MATRIX_NODE_DIAGNOSTICS_CSV_NAME,
-    REFRACTION_DESIGN_MATRIX_QC_JSON_NAME,
-)
 from app.services.refraction_static_export_units import (
     REFRACTION_STATIC_REPO_SIGN_CONVENTION,
-)
-from app.services.refraction_static_preflight_diagnostics import (
-    REFRACTION_STATIC_PREFLIGHT_OBSERVATIONS_CSV_NAME,
-    REFRACTION_STATIC_PREFLIGHT_QC_JSON_NAME,
-)
-from app.services.refraction_static_source_depth import (
-    REFRACTION_SOURCE_DEPTH_QC_JSON_NAME,
-    REFRACTION_SOURCE_DEPTH_SOURCES_CSV_NAME,
-)
-from app.services.refraction_static_t1lsst import (
-    REFRACTION_T1LSST_1LAYER_COMPONENTS_CSV_NAME,
 )
 from app.services.refraction_static_types import (
     RefractionDatumStaticsResult,
     RefractionLayerKind,
-)
-from app.services.refraction_static_uphole import (
-    REFRACTION_UPHOLE_QC_JSON_NAME,
-    REFRACTION_UPHOLE_SOURCES_CSV_NAME,
-)
-from app.services.refraction_static_v1 import (
-    REFRACTION_V1_ESTIMATES_CSV_NAME,
-    REFRACTION_V1_QC_JSON_NAME,
 )
 
 REFRACTION_STATIC_SOLUTION_NPZ_NAME = 'refraction_static_solution.npz'
@@ -129,84 +106,6 @@ class _CellVelocityArtifactPaths:
     solver_history_csv: Path
 
 
-_CELL_VELOCITY_COMPONENT_BY_LAYER: dict[RefractionLayerKind, str] = {
-    'v2_t1': 'v2',
-    'v3_t2': 'v3',
-    'vsub_t3': 'vsub',
-}
-_CELL_VELOCITY_LABEL_BY_LAYER: dict[RefractionLayerKind, str] = {
-    'v2_t1': 'V2/T1',
-    'v3_t2': 'V3/T2',
-    'vsub_t3': 'Vsub/T3',
-}
-_CELL_VELOCITY_ARTIFACT_NAMES_BY_LAYER: dict[
-    RefractionLayerKind,
-    _CellVelocityArtifactNames,
-] = {
-    'v2_t1': _CellVelocityArtifactNames(
-        cells_csv=REFRACTION_REFRACTOR_VELOCITY_CELLS_CSV_NAME,
-        grid_npz=REFRACTION_REFRACTOR_VELOCITY_GRID_NPZ_NAME,
-        qc_json=REFRACTION_REFRACTOR_VELOCITY_QC_JSON_NAME,
-        solver_history_csv=REFRACTION_CELL_SOLVER_HISTORY_CSV_NAME,
-    ),
-    'v3_t2': _CellVelocityArtifactNames(
-        cells_csv=REFRACTION_V3_REFRACTOR_VELOCITY_CELLS_CSV_NAME,
-        grid_npz=REFRACTION_V3_REFRACTOR_VELOCITY_GRID_NPZ_NAME,
-        qc_json=REFRACTION_V3_REFRACTOR_VELOCITY_QC_JSON_NAME,
-        solver_history_csv=REFRACTION_V3_CELL_SOLVER_HISTORY_CSV_NAME,
-    ),
-    'vsub_t3': _CellVelocityArtifactNames(
-        cells_csv=REFRACTION_VSUB_REFRACTOR_VELOCITY_CELLS_CSV_NAME,
-        grid_npz=REFRACTION_VSUB_REFRACTOR_VELOCITY_GRID_NPZ_NAME,
-        qc_json=REFRACTION_VSUB_REFRACTOR_VELOCITY_QC_JSON_NAME,
-        solver_history_csv=REFRACTION_VSUB_CELL_SOLVER_HISTORY_CSV_NAME,
-    ),
-}
-
-
-def _cell_velocity_artifact_names(
-    layer_kind: RefractionLayerKind,
-) -> _CellVelocityArtifactNames:
-    try:
-        return _CELL_VELOCITY_ARTIFACT_NAMES_BY_LAYER[layer_kind]
-    except KeyError as exc:
-        raise RefractionStaticArtifactError(
-            f'unsupported cell velocity layer kind: {layer_kind}'
-        ) from exc
-
-
-def _cell_velocity_artifact_entries_for_layer(
-    layer_kind: RefractionLayerKind,
-) -> tuple[dict[str, str | bool], ...]:
-    names = _cell_velocity_artifact_names(layer_kind)
-    label = _CELL_VELOCITY_LABEL_BY_LAYER[layer_kind]
-    return (
-        {
-            'name': names.cells_csv,
-            'kind': 'csv',
-            'required': True,
-            'description': f'Per-cell {label} refractor velocity grid and QC metrics',
-        },
-        {
-            'name': names.grid_npz,
-            'kind': 'npz',
-            'required': True,
-            'description': f'Machine-readable {label} refractor velocity cell grid',
-        },
-        {
-            'name': names.qc_json,
-            'kind': 'json',
-            'required': True,
-            'description': f'{label} refractor velocity cell QC summary',
-        },
-        {
-            'name': names.solver_history_csv,
-            'kind': 'csv',
-            'required': True,
-            'description': f'Cell {label} solver convergence and history summary',
-        },
-    )
-
 ARTIFACT_VERSION = '1.0'
 METHOD = 'gli_variable_thickness'
 WORKFLOW = 'refraction_statics'
@@ -223,306 +122,6 @@ FIRST_BREAK_TIME_EXPORT_FORMAT_VERSION = 1
 FIRST_BREAK_TIME_EXPORT_SIGN_CONVENTION = (
     'residual_ms = observed_pick_time_ms - modeled_pick_time_ms'
 )
-
-_ARTIFACTS: tuple[dict[str, str | bool], ...] = (
-    {
-        'name': REFRACTION_STATIC_SOLUTION_NPZ_NAME,
-        'kind': 'npz',
-        'required': True,
-        'description': 'Machine-readable final refraction statics solution',
-    },
-    {
-        'name': REFRACTION_STATIC_QC_JSON_NAME,
-        'kind': 'json',
-        'required': True,
-        'description': 'Human-readable final refraction statics QC summary',
-    },
-    {
-        'name': REFRACTION_STATIC_HISTORY_JSON_NAME,
-        'kind': 'json',
-        'required': True,
-        'description': 'Static-component lineage and double-application audit history',
-    },
-    {
-        'name': REFRACTION_STATICS_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Trace-level final refraction statics table',
-    },
-    {
-        'name': NEAR_SURFACE_MODEL_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Node-level near-surface model table',
-    },
-    {
-        'name': FIRST_BREAK_RESIDUALS_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'GLI first-break residual table',
-    },
-    {
-        'name': REFRACTION_FIRST_BREAK_TIME_EXPORT_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Observation-level first-break time QC export',
-    },
-    {
-        'name': REFRACTION_FIRST_BREAK_FIT_QC_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Viewer-ready observed-modeled first-break fit QC table',
-    },
-    {
-        'name': REFRACTION_FIRST_BREAK_FIT_QC_NPZ_NAME,
-        'kind': 'npz',
-        'required': True,
-        'description': 'Machine-readable observed-modeled first-break fit QC arrays',
-    },
-    {
-        'name': REFRACTION_FIRST_BREAK_FIT_QC_JSON_NAME,
-        'kind': 'json',
-        'required': True,
-        'description': 'Observed-modeled first-break fit QC schema and summary',
-    },
-    {
-        'name': REFRACTION_REDUCED_TIME_QC_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Reduced-time first-break QC table for LMO displays',
-    },
-    {
-        'name': REFRACTION_REDUCED_TIME_QC_NPZ_NAME,
-        'kind': 'npz',
-        'required': True,
-        'description': 'Machine-readable reduced-time first-break QC arrays',
-    },
-    {
-        'name': REFRACTION_REDUCED_TIME_QC_JSON_NAME,
-        'kind': 'json',
-        'required': True,
-        'description': 'Reduced-time first-break QC schema and summary',
-    },
-    {
-        'name': REFRACTION_STATIC_COMPONENTS_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Source/receiver endpoint static component table',
-    },
-    {
-        'name': REFRACTION_STATIC_COMPONENT_QC_TRACE_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Trace-level static component waterfall QC table',
-    },
-    {
-        'name': REFRACTION_STATIC_COMPONENT_QC_ENDPOINT_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Endpoint-level static component waterfall QC table',
-    },
-    {
-        'name': REFRACTION_STATIC_COMPONENT_QC_NPZ_NAME,
-        'kind': 'npz',
-        'required': True,
-        'description': 'Machine-readable static component waterfall QC arrays',
-    },
-    {
-        'name': REFRACTION_STATIC_COMPONENT_QC_JSON_NAME,
-        'kind': 'json',
-        'required': True,
-        'description': 'Static component waterfall QC schema and summary',
-    },
-    {
-        'name': SOURCE_STATIC_TABLE_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'IRAS-style source endpoint final static table',
-    },
-    {
-        'name': RECEIVER_STATIC_TABLE_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'IRAS-style receiver endpoint final static table',
-    },
-    {
-        'name': SOURCE_RECEIVER_STATIC_TABLE_NPZ_NAME,
-        'kind': 'npz',
-        'required': True,
-        'description': 'Machine-readable source/receiver endpoint static tables',
-    },
-    {
-        'name': REFRACTION_LINE_PROFILE_QC_SOURCE_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Source endpoint line-profile QC rows sorted by inline distance',
-    },
-    {
-        'name': REFRACTION_LINE_PROFILE_QC_RECEIVER_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Receiver endpoint line-profile QC rows sorted by inline distance',
-    },
-    {
-        'name': REFRACTION_LINE_PROFILE_QC_COMBINED_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Combined source/receiver line-profile QC rows',
-    },
-    {
-        'name': REFRACTION_LINE_PROFILE_QC_NPZ_NAME,
-        'kind': 'npz',
-        'required': True,
-        'description': 'Machine-readable source/receiver line-profile QC arrays',
-    },
-    {
-        'name': REFRACTION_LINE_PROFILE_QC_JSON_NAME,
-        'kind': 'json',
-        'required': True,
-        'description': 'Line-profile QC schema, availability, and summary',
-    },
-    {
-        'name': REFRACTION_TIME_TERM_SPREADSHEET_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Spreadsheet endpoint time terms, layers, statics, and statuses',
-    },
-)
-
-_V1_ARTIFACTS: tuple[dict[str, str | bool], ...] = (
-    {
-        'name': REFRACTION_V1_QC_JSON_NAME,
-        'kind': 'json',
-        'required': True,
-        'origin': 'upstream',
-        'description': 'Direct-arrival V1 estimation QC summary',
-    },
-    {
-        'name': REFRACTION_V1_ESTIMATES_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'origin': 'upstream',
-        'description': 'Per-source direct-arrival V1 estimates',
-    },
-)
-
-_V1_ARTIFACT_NAMES = frozenset(
-    {
-        REFRACTION_V1_QC_JSON_NAME,
-        REFRACTION_V1_ESTIMATES_CSV_NAME,
-    }
-)
-
-_SOURCE_DEPTH_FIELD_ARTIFACTS: tuple[dict[str, str | bool], ...] = (
-    {
-        'name': REFRACTION_SOURCE_DEPTH_QC_JSON_NAME,
-        'kind': 'json',
-        'required': True,
-        'origin': 'upstream',
-        'description': 'Source-depth field-correction QC summary',
-    },
-    {
-        'name': REFRACTION_SOURCE_DEPTH_SOURCES_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'origin': 'upstream',
-        'description': 'Resolved source-depth rows used by field corrections',
-    },
-)
-_SOURCE_DEPTH_FIELD_ARTIFACT_NAMES = frozenset(
-    str(item['name']) for item in _SOURCE_DEPTH_FIELD_ARTIFACTS
-)
-
-_UPHOLE_FIELD_ARTIFACTS: tuple[dict[str, str | bool], ...] = (
-    {
-        'name': REFRACTION_UPHOLE_QC_JSON_NAME,
-        'kind': 'json',
-        'required': True,
-        'origin': 'upstream',
-        'description': 'Uphole-time field-correction QC summary',
-    },
-    {
-        'name': REFRACTION_UPHOLE_SOURCES_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'origin': 'upstream',
-        'description': 'Resolved source uphole-time rows used by field corrections',
-    },
-)
-_UPHOLE_FIELD_ARTIFACT_NAMES = frozenset(
-    str(item['name']) for item in _UPHOLE_FIELD_ARTIFACTS
-)
-
-_UPSTREAM_ARTIFACTS: tuple[dict[str, str | bool], ...] = (
-    _V1_ARTIFACTS + _SOURCE_DEPTH_FIELD_ARTIFACTS + _UPHOLE_FIELD_ARTIFACTS
-)
-_UPSTREAM_ARTIFACT_NAMES = frozenset(
-    str(item['name']) for item in _UPSTREAM_ARTIFACTS
-)
-
-_T1LSST_1LAYER_ARTIFACTS: tuple[dict[str, str | bool], ...] = (
-    {
-        'name': REFRACTION_T1LSST_1LAYER_COMPONENTS_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'T1LSST-compatible one-layer source/receiver components',
-    },
-)
-
-_REFRACTOR_CELL_VELOCITY_ARTIFACTS = _cell_velocity_artifact_entries_for_layer(
-    'v2_t1'
-)
-_ALL_REFRACTOR_CELL_VELOCITY_ARTIFACTS: tuple[dict[str, str | bool], ...] = (
-    _REFRACTOR_CELL_VELOCITY_ARTIFACTS
-    + _cell_velocity_artifact_entries_for_layer('v3_t2')
-    + _cell_velocity_artifact_entries_for_layer('vsub_t3')
-)
-_GRID_MAP_QC_ARTIFACTS: tuple[dict[str, str | bool], ...] = (
-    {
-        'name': REFRACTION_GRID_MAP_QC_CSV_NAME,
-        'kind': 'csv',
-        'required': True,
-        'description': 'Viewer-ready refraction cell velocity grid map QC rows',
-    },
-    {
-        'name': REFRACTION_GRID_MAP_QC_NPZ_NAME,
-        'kind': 'npz',
-        'required': True,
-        'description': 'Machine-readable refraction cell velocity grid map QC arrays',
-    },
-    {
-        'name': REFRACTION_GRID_MAP_QC_JSON_NAME,
-        'kind': 'json',
-        'required': True,
-        'description': 'Refraction cell velocity grid map QC summary',
-    },
-)
-
-_ARTIFACT_CONTENT_TYPE_BY_KIND = {
-    'csv': 'text/csv',
-    'json': 'application/json',
-    'npz': 'application/octet-stream',
-}
-
-REFRACTION_STATIC_REGISTERED_ARTIFACT_NAMES = frozenset(
-    str(item['name'])
-    for item in (
-        _ARTIFACTS
-        + _UPSTREAM_ARTIFACTS
-        + _T1LSST_1LAYER_ARTIFACTS
-        + _ALL_REFRACTOR_CELL_VELOCITY_ARTIFACTS
-        + _GRID_MAP_QC_ARTIFACTS
-    )
-) | {
-    REFRACTION_STATIC_ARTIFACTS_JSON_NAME,
-    REFRACTION_STATIC_REQUEST_JSON_NAME,
-    REFRACTION_STATIC_FAILURE_DIAGNOSTICS_JSON_NAME,
-    REFRACTION_STATIC_PREFLIGHT_QC_JSON_NAME,
-    REFRACTION_STATIC_PREFLIGHT_OBSERVATIONS_CSV_NAME,
-    REFRACTION_DESIGN_MATRIX_QC_JSON_NAME,
-    REFRACTION_DESIGN_MATRIX_NODE_DIAGNOSTICS_CSV_NAME,
-    UPLOADED_REFRACTION_PICKS_NPZ_NAME,
-}
 
 _TRACE_STATICS_COLUMNS = (
     'sorted_trace_index',
@@ -1345,7 +944,6 @@ __all__ = [
     'REFRACTION_STATIC_FAILURE_DIAGNOSTICS_JSON_NAME',
     'REFRACTION_STATIC_HISTORY_JSON_NAME',
     'REFRACTION_STATIC_QC_JSON_NAME',
-    'REFRACTION_STATIC_REGISTERED_ARTIFACT_NAMES',
     'REFRACTION_STATIC_REQUEST_JSON_NAME',
     'REFRACTION_STATIC_SOLUTION_NPZ_NAME',
     'REFRACTION_STATICS_CSV_NAME',
